column no should be an int
git-svn-id: http://svn.xp-dev.com/svn/wxextension@2700 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/include/wx/extension/listview.h
@@ -131,7 +131,7 @@ public:
   wxExImageType GetImageType() const {return m_ImageType;};
 
   /// Gets the item text using item number and column number.
-  const wxString GetItemText(long item_number, long col_number) const;
+  const wxString GetItemText(long item_number, int col_number) const;
 
   /// Gets the item text using item number and column name.
   const wxString GetItemText(
